backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-community
langchain-core
langchain-openai
langchain-groq
python-dotenv
pydantic
requests
beautifulsoup4
pdfplumber
chromadb
sentence-transformers
rank_bm25
pandas
openpyxl
websockets
python-multipart
redis
feedparser
sentry-sdk[fastapi]
PyPDF2
pymupdf
python-docx
# OCR & NLP enhancements
easyocr
Pillow
pdf2image
spacy
scikit-learn
numpy
